back/requirements.txt
asgiref==3.8.1
class-registry==2.1.2
Django==5.0.4
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==24.2
django-js-asset==2.2.0
djangorestframework==3.15.1
filters==1.3.2
filters-django==1.0.5
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.4.28
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
